Refactor engines to have a base class
